refactor: Post의 Comment 코드 분리
